+++ src/ndtm.cpp
@@ -38,7 +38,7 @@ void NDTM::loadTapes(std::shared_ptr<Tape> t) {
 
 std::string NDTM::readSymbols() {
     std::string s;
-    for (int i = 0; i < m_Tapes.size(); i++) {
+    for (int i = 0; i < (int) m_Tapes.size(); i++) {
         s.push_back(m_Tapes[i]->readSymbol());
     }
     return s;
@@ -52,7 +52,7 @@ void NDTM::execTransition(std::pair<int, tapesOperations> trans) {
     m_CurState = trans.first;
 
 
-    for (int i = 0; i < trans.second.size(); i++) {
+    for (int i = 0; i < (int) trans.second.size(); i++) {
         if (trans.second[i].first != 'E' && trans.second[i].first != 'A')
             m_Tapes[i]->writeSymbol(trans.second[i].first);
         m_Tapes[i]->moveHead(trans.second[i].second);
@@ -62,7 +62,7 @@ void NDTM::execTransition(std::pair<int, tapesOperations> trans) {
 bool NDTM::accepts() {
     auto trans = m_Transitions[std::make_pair(m_CurState, this->readSymbols())];
     if (this->checkCycle()) return false;
-    for (int i = 0; i < trans.size(); i++) {
+    for (int i = 0; i < (int) trans.size(); i++) {
         auto tm = this->clone();
         tm->execTransition(trans[i]);
         if (tm->accepts()) return true;
@@ -103,14 +103,14 @@ void NDTM::print() {
     std::cout << "Transition function f:" << std::endl;
     for (auto it = m_Transitions.begin(); it != m_Transitions.end(); it++) {
         std::cout << "f(" << it->first.first;
-        for (int i = 0; i < m_Transitions.begin()->first.second.size(); i++) {
+        for (int i = 0; i < (int) m_Transitions.begin()->first.second.size(); i++) {
             if (it->first.second[i] == m_Blank) std::cout << ",B";
             else std::cout << "," << it->first.second[i];
         }
         std::cout << ") = {";
         for (auto it1 = it->second.begin(); it1 != it->second.end(); it1++) {
             std::cout << "\n\t(" << it1->first;
-            for (int i = 0; i < m_Transitions.begin()->first.second.size(); i++) {
+            for (int i = 0; i < (int) m_Transitions.begin()->first.second.size(); i++) {
                 if (it1->second[i].first == m_Blank) std::cout << ",<B," << it1->second[i].second - 1 << ">";
                 else std::cout << ",<" << it1->second[i].first << "," << it1->second[i].second - 1 << ">";
             }
@@ -128,14 +128,14 @@ void NDTM::initialize(std::string input) {
 bool NDTM::checkCycle() {
     std::string s = "";
     std::vector<int> pos;
-    for (int i = 0; i < m_Tapes.size(); i++) {
+    for (int i = 0; i < (int) m_Tapes.size(); i++) {
         s += getTape(i)->getMCells();
         pos.push_back(getTape(i)->getMHead());
     }
 
     if (!m_ConfigurationsMemory[m_CurState].first.empty()) {
         bool hasCycle = true;
-        for (int i = 0; i < m_Tapes.size(); i++) {
+        for (int i = 0; i < (int) m_Tapes.size(); i++) {
             if (m_ConfigurationsMemory[m_CurState].second[i] != getTape(i)->getMHead()) hasCycle = false;
         }
         if (m_ConfigurationsMemory[m_CurState].first == s && hasCycle) return true;
@@ -156,7 +156,7 @@ void NDTM::epsilonTransitionHelper(std::vector<std::string> &readSymbols, std::v
         }
         auto it = alphabet.begin();
         int size = readSymbols.size();
-        for (int i = 0; i < alphabet.size() * size; i++) {
+        for (int i = 0; i < (int) alphabet.size() * size; i++) {
             if (i >= size) {
                 std::string s = readSymbols[i % size];
                 tapesOperations t = operations[i % size];
